SECURITY FIX: removed mpdf example scripts and added .htaccess
The mpdf library used in this plugin has a vulnerability in one of the
example scripts: http://www.mpdf1.com/mpdf/forum/comments.php?DiscussionID=781

This patch removes the whole example directory as it isn't needed
anyway. It also adds a .htaccess file to deny access to the library
folder from the web.