This is a quick an dirty switch to the current MPDF devel version
This makes the plugin working on PHP 7.1 again, however I did not look
deeply in how much the API changed and just tried to quickly make the
whole thing work again.

I opted for a composer based install of the library to have the proper
dependencies installed. I also checked in the result to have the plugin
easily install via the extension manager.