fix: re-enable caching for ns export edge case
If a namespace is exported while viewing an old revision or the wiki in
an old version, the current version of the namespace is exported. This
is correct, However it makes sense to chache this file or use the cache
if it exists and this is now re-enabled after it was disabled in a
regression in the last commit.