fix: print ACL'd images again
The update of the upstream mpdf library in e56751d undid the adjustment
made for the images in 0576bb8.

Since the approach chosen in that commit was declined upstream
(mpdf/mpdf#460), this is now a minimal fix for that problem until we
find time to fix it properly upstream.

This fix will have to be repeated after every update of the upstream
library.