Fix multi page rendering by setting global $INFO
In case of rendering multiple pages into a single document, the
context of the current page does not match the one of the pages
to be rendered. In some cases (e.g. with the struct plugin), the
global $INFO is evluated and as this is not set to match the page
to be rendered, the struct plugin does not produce the expected
output in the PDF.

Fix this by setting $INFO to match the current page to be rendered
and restoring it afterwards.

Signed-off-by: Frieder Schrempf <dev@fris.de>