update MPDF downgrade log/psr
mpdf v8.2.6 => v8.2.7
log 3.0.2 => 1.1.4

Using the old log version avoids conflicts with the smtp plugin for now.

However we need a proper solution to this problem in the future
(probably in core).

A PR to be able to use any log version with the smtp plugin has been
submitted: https://github.com/txthinking/Mailer/pull/36

We probably should introduce a log/psr:3 dependency in core when
upgrading to PHP8 (and provide a PSR compatible wrapper around our
logger)